+++ tests/test_iqm_job.py
@@ -16,7 +16,7 @@
 """
 import uuid
 
-from iqm_client import IQMClient, RunResult, RunStatus, Status
+from iqm_client import HeraldingMode, Instruction, IQMClient, RunResult, RunStatus, SingleQubitMapping, Status
 import mockito
 from mockito import mock, when
 import pytest
@@ -37,8 +37,8 @@ def job(adonis_architecture):
 
 
 @pytest.fixture()
-def iqm_result_single_register():
-    return {'c_2_0_0': [[0], [1], [0], [1]], 'c_2_0_1': [[1], [1], [1], [0]]}
+def iqm_result_no_shots():
+    return {'c_2_0_0': [], 'c_2_0_1': []}
 
 
 @pytest.fixture()
@@ -48,12 +48,17 @@ def iqm_result_two_registers():
 
 @pytest.fixture()
 def iqm_metadata():
+    measurement = Instruction(name='measurement', implementation=None, qubits=('0',), args={'key': 'm1'})
     return {
         'calibration_set_id': 'df124054-f6d8-41f9-b880-8487f90018f9',
         'request': {
             'shots': 4,
-            'circuits': [{'name': 'circuit_1', 'instructions': [], 'metadata': {'a': 'b'}}],
+            'circuits': [{'name': 'circuit_1', 'instructions': (measurement,), 'metadata': {'a': 'b'}}],
             'calibration_set_id': 'df124054-f6d8-41f9-b880-8487f90018f9',
+            'qubit_mapping': [
+                SingleQubitMapping(logical_name='0', physical_name='QB1'),
+                SingleQubitMapping(logical_name='1', physical_name='QB2'),
+            ],
         },
     }
 
@@ -83,8 +88,9 @@ def test_status_done(job, iqm_metadata):
     assert job._result is None
 
 
-def test_status_running(job):
-    when(job._client).get_run_status(uuid.UUID(job.job_id())).thenReturn(RunStatus(status=Status.PENDING))
+@pytest.mark.parametrize('run_status', [Status.PENDING_COMPILATION, Status.PENDING_EXECUTION])
+def test_status_running(job, run_status):
+    when(job._client).get_run_status(uuid.UUID(job.job_id())).thenReturn(RunStatus(status=run_status))
     assert job.status() == JobStatus.RUNNING
 
 
@@ -109,6 +115,7 @@ def test_result(job, iqm_result_two_registers, iqm_metadata):
     for r in result.results:
         assert r.calibration_set_id == uuid.UUID('df124054-f6d8-41f9-b880-8487f90018f9')
         assert r.data.metadata == {'a': 'b'}
+    assert result.request.qubit_mapping == iqm_metadata['request']['qubit_mapping']
 
     # Assert that repeated call does not query the client (i.e. works without calling the mocked wait_for_results)
     # and call to status() does not call any functions from client.
@@ -118,16 +125,39 @@ def test_result(job, iqm_result_two_registers, iqm_metadata):
     mockito.verify(job._client, times=1).wait_for_results(uuid.UUID(job.job_id()))
 
 
+def test_result_no_shots(job, iqm_result_no_shots, iqm_metadata):
+    iqm_metadata['request']['heralding_mode'] = HeraldingMode.ZEROS
+    client_result = RunResult(
+        status=Status.READY,
+        measurements=[iqm_result_no_shots],
+        metadata=iqm_metadata,
+    )
+    when(job._client).wait_for_results(uuid.UUID(job.job_id())).thenReturn(client_result)
+
+    with pytest.warns(UserWarning, match='Received measurement results containing zero shots.'):
+        result = job.result()
+
+    assert isinstance(result, QiskitResult)
+    assert result.get_memory() == []
+    assert result.get_counts() == Counts({})
+
+
 def test_result_multiple_circuits(job, iqm_result_two_registers):
+    instruction_meta = [{'name': 'measurement', 'qubits': ['0'], 'args': {'key': 'm1'}}]
     iqm_metadata_multiple_circuits = {
         'calibration_set_id': '9d75904b-0c93-461f-b1dc-bd200cfad1f1',
         'request': {
             'shots': 4,
             'circuits': [
-                {'name': 'circuit_1', 'instructions': [], 'metadata': {'a': 0}},
-                {'name': 'circuit_2', 'instructions': [], 'metadata': {'a': 1}},
+                {'name': 'circuit_1', 'instructions': instruction_meta, 'metadata': {'a': 0}},
+                {'name': 'circuit_2', 'instructions': instruction_meta, 'metadata': {'a': 1}},
             ],
             'calibration_set_id': '9d75904b-0c93-461f-b1dc-bd200cfad1f1',
+            'qubit_mapping': [
+                SingleQubitMapping(logical_name='0', physical_name='QB1'),
+                SingleQubitMapping(logical_name='1', physical_name='QB2'),
+                SingleQubitMapping(logical_name='2', physical_name='QB3'),
+            ],
         },
     }
     client_result = RunResult(
@@ -148,3 +178,4 @@ def test_result_multiple_circuits(job, iqm_result_two_registers):
     for i, r in enumerate(result.results):
         assert r.calibration_set_id == uuid.UUID('9d75904b-0c93-461f-b1dc-bd200cfad1f1')
         assert r.data.metadata == {'a': i}
+    assert result.request.qubit_mapping == iqm_metadata_multiple_circuits['request']['qubit_mapping']